[MDEPLOY-265] Allow old alt*DeploymentRepository property format if default layout is used
The legacy format (<=2.x) of alt*DeploymentRepository is id::layout::url.
The new format (>= 3.x) of alt*DeploymentRepository is id::url which is
equivalent to id::default:url from the legacy format.

This change introduces backwards compatibility with 2.x by supporting
alt*DeploymentRepository values in the id::layout::url format if and only if
the layout is equal to "default".  The "default" layout is the most commonly
used layout, so this should maintain backwards compatibility with the large
majority of projects using the alt*DeploymentRepository properties.

* Usage of the legacy format with the "default" layout will result in a warning
  message being logged.
* Usage of the legacy format with layouts other than "default" will result in
  an exception being thrown.

This closes #15