[MDEPLOY-265] Maintain backwards compatibility for alt*DeploymentRepository in id::default::url format
The legacy format (<=2.x) of alt*DeploymentRepository is id::layout::url.
The new format (>= 3.x) of alt*DeploymentRepository is id::url, which is equivalent to id::default:url from the legacy format.

This change introduces backwards compatibility with 2.x by supporting alt*DeploymentRepository values in the id::layout::url format if and only if the layout is equal to "default".  The "default" layout is the most commonly used layout, so this should maintain backwards compatibility with the large majority of projects using the alt*DeploymentRepository properties.

Usage of the legacy format with the "default" layout will result in a warning message being logged.
Usage of the legacy format with layouts other than "default" will result in an exception being thrown.

One use case this change will help is when alt*DeploymentRepository properties are set in common maven settings.xml files used for a large number of projects.  For example, a CI server.  Keeping backwards compatibilty with id::default::url allows those settings.xml files to be used regardless of the version of maven-deploy-plugin used by the projects built on that CI server.  Projects can individually upgrade their maven-deploy-plugin without breaking.