[MDEPLOY-291] Update POM parent and Maven
Update plugin build:
* update parent to 36
* update maven to 3.8.5 (reasoning below)
* up the plugin to Java8

This plugin along with m-install-p versions 3.x line (3.0.0-Mx) does not work with any so far released Maven 3.x line (see MNG-7055). The ONLY maven 3.x they will be usable with is upcoming 3.9.0, that has fix for MNG-7055 backported, but OTOH, it is already Java8, so no point in keeping the plugin on anything below.

---

https://issues.apache.org/jira/browse/MDEPLOY-291